backend/requirements.txt
# Web framework
django==5.0.9
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0

# Auth
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Cache / Task queue
celery[redis]==5.4.0
django-celery-beat==2.7.0
django-redis==5.4.0
redis==5.0.8

# Storage (MinIO / S3)
django-storages[s3]==1.14.4
boto3==1.35.0

# Vedic astrology
pyswisseph==2.10.3.2

# Geocoding & timezone
geopy==2.4.1
timezonefinder==6.5.2

# LLM (Ollama via HTTP)
httpx==0.27.2

# Image processing
Pillow==10.4.0

# Config
python-decouple==3.8

# WSGI server
gunicorn==22.0.0

# Static file serving
whitenoise==6.7.0
